requirements.txt
pytest-homeassistant-custom-component~=0.13.134
flake8~=5.0.4
pytest~=8.2.0
python-frank-energie~=6.0.0
